Fixes #3347 (backported from next 946c1777)

+++ companion/src/firmwares/opentx/opentxeeprom.cpp
@@ -1976,7 +1976,7 @@ class ArmCustomFunctionField: public TransformedField {
 
       if (fn.func >= FuncOverrideCH1 && fn.func <= FuncOverrideCH32) {
         fn.func = AssignFunc(fn.func + index);
-        fn.param = (int8_t)value;
+        fn.param = (int16_t)(uint16_t)value;
       }
       else if (fn.func >= FuncSetTimer1 && fn.func <= FuncSetTimer2) {
         fn.func = AssignFunc(fn.func + index);
